test: Helmet e2e
