+++ packages/playground/server-components/src/pages/stream.server.jsx
@@ -1,32 +1,6 @@
-import {defer} from '@shopify/hydrogen/dist/esnext/utilities/defer';
 import {Suspense} from 'react';
 import Counter from '../components/Counter.client';
-import {useServerRequest} from '@shopify/hydrogen/dist/esnext/foundation/ServerRequestProvider';
-
-const createData = (key, ms) => {
-  const {cache} = useServerRequest().ctx;
-
-  if (!cache.has(key)) {
-    let result;
-    const deferred = defer();
-    setTimeout(deferred.resolve, ms);
-    deferred.promise.then((r) => {
-      result = r;
-    });
-
-    cache.set(key, {
-      read: () => {
-        if (deferred.status === 'pending') {
-          throw deferred.promise;
-        }
-
-        return result || 'done';
-      },
-    });
-  }
-
-  return cache.get(key);
-};
+import {createData} from '../utils';
 
 // Page with many Suspense boundaries to test streaming
 export default function Index() {
